fix: moved logic to model
