+++ src/renderer/features/multisig-operations/model/approve-model.ts
@@ -4,7 +4,7 @@ import { combine, createEffect, createEvent, createStore, restore, sample } from
 import { createGate } from 'effector-react';
 
 import { type Chain } from '@/shared/core';
-import { getNativeAsset, nonNullable, nullable, validateCallData } from '@/shared/lib/utils';
+import { getNativeAsset, groupBy, nonNullable, nullable, validateCallData } from '@/shared/lib/utils';
 import {
   createComplexTxStore,
   createMultisigDeposit,
@@ -22,6 +22,7 @@ import {
 import { balanceModel } from '@/entities/balance';
 import { networkModel } from '@/entities/network';
 import { MAX_WEIGHT, getExtrinsic, transactionBuilder } from '@/entities/transaction';
+import { accountUtils } from '@/entities/wallet';
 
 import { operationsContextModel } from './context';
 
@@ -57,6 +58,49 @@ const $api = combine(
   },
 );
 
+const $unsignedAccounts = combine(
+  {
+    multisigAccount: operationsContextModel.$multisigAccount,
+    chain: $chain,
+    accountsList: accounts.$list,
+    operation: $operation,
+  },
+  ({ multisigAccount, chain, accountsList, operation }) => {
+    console.log('ya ebal kozu unsignedAccounts', { multisigAccount, chain, accountsList, operation });
+    if (!multisigAccount || !chain || !operation) return [];
+
+    const signatories = accountsList.filter(a =>
+      multisigAccount.signatories.some(s => s.accountId === a.accountId && (s.id ? s.id === a.walletId : true)),
+    );
+
+    const signatoriesOnChain = signatories.filter(s => (s.type === 'chain' ? s.chainId === chain.chainId : true));
+
+    const filteredSignatories = signatoriesOnChain.filter(
+      a => !operation.events.some(e => e.accountId === a.accountId),
+    );
+
+    const signatoriesGroupedByAccountId = groupBy(filteredSignatories, a => a.accountId);
+
+    let result: AnyAccount[] = [];
+
+    for (const group of Object.values(signatoriesGroupedByAccountId)) {
+      if (nullable(group)) continue;
+
+      if (group.length > 1) {
+        const flexibleMultisigAccount = group.find(accountUtils.isFlexibleMultisigAccount);
+        if (flexibleMultisigAccount) {
+          result.push(flexibleMultisigAccount);
+          continue;
+        }
+      }
+
+      result = result.concat(group);
+    }
+
+    return result;
+  },
+);
+
 const $signatories = createSignatoriesStore({
   chain: $chain,
   initiator: $initiator,
@@ -64,9 +108,15 @@ const $signatories = createSignatoriesStore({
 });
 
 sample({
+  clock: $unsignedAccounts,
+  filter: $unsignedAccounts.map(unsignedAccounts => unsignedAccounts.length === 1),
+  fn: unsignedAccounts => unsignedAccounts.at(0) ?? null,
+  target: $initiator,
+});
+sample({
   clock: $signatories,
   filter: $signatories.map(signatories => signatories.length === 1),
-  fn: signatories => signatories[0],
+  fn: signatories => signatories.at(0) ?? null,
   target: $signatory,
 });
 
@@ -195,6 +245,7 @@ export const approveModel = {
   $signatory,
   $signingPayloads,
   $initiator,
+  $unsignedAccounts,
 
   $signatories,
   selectSignatory,
